PIOS: Remove a lot of FreeRTOS
