+++ flight/PiOS/STM32F10x/pios_usb_cdc.c
@@ -34,10 +34,6 @@
 
 #if defined(PIOS_INCLUDE_USB_CDC)
 
-#if defined(PIOS_INCLUDE_FREERTOS)
-#include "FreeRTOS.h"
-#endif /* defined(PIOS_INCLUDE_FREERTOS) */
-
 #include "pios_usb.h"
 #include "pios_usb_cdc_priv.h"
 #include "pios_usb_board_data.h" /* PIOS_BOARD_*_DATA_LENGTH */
@@ -216,10 +212,6 @@ static void PIOS_USB_CDC_SendData(struct pios_usb_cdc_dev * usb_cdc_dev)
 			bytes_to_tx);
 	SetEPTxCount(usb_cdc_dev->cfg->data_tx_ep, bytes_to_tx);
 	SetEPTxValid(usb_cdc_dev->cfg->data_tx_ep);
-
-#if defined(PIOS_INCLUDE_FREERTOS)
-	portEND_SWITCHING_ISR(need_yield ? pdTRUE : pdFALSE);
-#endif	/* PIOS_INCLUDE_FREERTOS */
 }
 
 static void PIOS_USB_CDC_TxStart(uintptr_t usbcdc_id, uint16_t tx_bytes_avail)
@@ -299,10 +291,6 @@ static void PIOS_USB_CDC_DATA_EP_OUT_Callback(void)
 		/* Not enough room left for a message, apply backpressure */
 		SetEPRxStatus(usb_cdc_dev->cfg->data_rx_ep, EP_RX_NAK);
 	}
-
-#if defined(PIOS_INCLUDE_FREERTOS)
-	portEND_SWITCHING_ISR(need_yield ? pdTRUE : pdFALSE);
-#endif	/* PIOS_INCLUDE_FREERTOS */
 }
 
 static uint16_t control_line_state;
